requirements.txt
# ValuePriorityBench Dependencies
# Core
python-dotenv>=1.0.0
numpy>=1.24.0
scipy>=1.11.0

# API Clients
anthropic>=0.18.0
openai>=1.12.0
google-generativeai>=0.4.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: Full Bayesian inference
# pymc>=5.10.0
# arviz>=0.17.0
